feat: refactor handlers to pkg (#44)
