+++ pkg/discovery/content/registry/mirror_test.go
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
-package v2
+package registry
 
 import (
 	"fmt"
